✒️ Rename: profile -> user로 수정 (#96)
- 폴더 수정
- 클래스 이름 수정
- 그로 인한 폴더 이동
