+++ src/main/kotlin/com/eatsfinder/domain/user/dto/user/active/MyActiveCommentResponse.kt
@@ -1,4 +1,4 @@
-package com.eatsfinder.domain.user.dto.profile.myactive
+package com.eatsfinder.domain.user.dto.user.active
 
 import java.time.LocalDateTime
