+++ src/main/kotlin/com/eatsfinder/domain/user/dto/user/active/MyActiveCommentUserResponse.kt
@@ -1,4 +1,4 @@
-package com.eatsfinder.domain.user.dto.profile.myactive
+package com.eatsfinder.domain.user.dto.user.active
 
 data class MyActiveCommentUserResponse(
     val commentUserNickname: String?,
