+++ src/main/kotlin/com/eatsfinder/domain/user/dto/user/active/MyActivePostLikeResponse.kt
@@ -1,4 +1,4 @@
-package com.eatsfinder.domain.user.dto.profile.myactive
+package com.eatsfinder.domain.user.dto.user.active
 
 data class MyActivePostLikeResponse (
     val postId: Long?,
